backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
geoalchemy2>=0.14.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
websockets>=12.0
redis>=5.0.0
celery>=5.3.0
httpx>=0.24.0
h3>=3.7.0
numpy>=1.26.0
onnxruntime>=1.16.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
pytz>=2024.1
